requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
psycopg2==2.8.3
python-dotenv==0.10.3
python-engineio==3.9.3
python-socketio==4.3.1
six==1.12.0
Werkzeug==0.15.5
